+++ src/App.tsx
@@ -1,102 +1,22 @@
-import { useEffect, useRef, useState } from 'react';
 import Background from './components/Background';
 import QuickButton from './components/QuickButton';
 import Section from './components/Section';
 import TopBar from './components/TopBar';
 import Nav from './components/Nav';
+import { useScrollStore } from './stores/scrollStore';
+import { useScrollHandler } from './hooks/useScrollHandler';
 
 function App() {
-  const [showTopBar, setShowTopBar] = useState(false);
-  const [isScrolling, setIsScrolling] = useState(false);
-
-  useEffect(() => {
-    let scrollEndTimer: number;
-    let isTouchDown = false;
-    let isMouseDown = false;
-    let scrollStartY = 0;
-    let hasReachedThreshold = false;
-
-    const pollForScrollEnd = () => {
-      // Only set scrolling to false if user is not touching/pressing
-      if (!isTouchDown && !isMouseDown) {
-        setIsScrolling(false);
-        hasReachedThreshold = false; // Reset threshold when scrolling ends
-      } else {
-        // Keep polling if user is still touching
-        scrollEndTimer = setTimeout(pollForScrollEnd, 50);
-      }
-    };
-
-    const handleScroll = () => {
-      if (window.scrollY > 10) {
-        setShowTopBar(true);
-      } else {
-        setShowTopBar(false);
-      }
-
-      // Check if we've scrolled 200px from start position
-      const currentScrollY = window.scrollY;
-      const scrollDistance = Math.abs(currentScrollY - scrollStartY);
-
-      if (!hasReachedThreshold && scrollDistance >= 50) {
-        hasReachedThreshold = true;
-      }
-
-      // Only set scrolling to true if we've reached the threshold
-      if (hasReachedThreshold) {
-        setIsScrolling(true);
-
-        // Clear existing timer
-        clearTimeout(scrollEndTimer);
-
-        // Start polling after 150ms of no scrolling
-        scrollEndTimer = setTimeout(pollForScrollEnd, 50);
-      }
-    };
-
-    const handleMouseDown = () => {
-      isMouseDown = true;
-      scrollStartY = window.scrollY; // Capture start position
-      clearTimeout(scrollEndTimer);
-    };
-
-    const handleMouseUp = () => {
-      isMouseDown = false;
-    };
-
-    const handleTouchStart = () => {
-      isTouchDown = true;
-      scrollStartY = window.scrollY; // Capture start position
-      clearTimeout(scrollEndTimer);
-    };
-
-    const handleTouchEnd = () => {
-      isTouchDown = false;
-    };
-
-    window.addEventListener('scroll', handleScroll);
-    window.addEventListener('mousedown', handleMouseDown);
-    window.addEventListener('mouseup', handleMouseUp);
-    window.addEventListener('touchstart', handleTouchStart);
-    window.addEventListener('touchend', handleTouchEnd);
-
-    return () => {
-      window.removeEventListener('scroll', handleScroll);
-      window.removeEventListener('mousedown', handleMouseDown);
-      window.removeEventListener('mouseup', handleMouseUp);
-      window.removeEventListener('touchstart', handleTouchStart);
-      window.removeEventListener('touchend', handleTouchEnd);
-      clearTimeout(scrollEndTimer);
-    };
-  }, []);
+  const { isScrolling, showTopBar } = useScrollStore();
+  useScrollHandler();
 
   return (
     <Background className={`font-courier pb-24`}>
       <Nav hide={isScrolling} />
       {/* top bar */}
-      {/* main content */}
       <TopBar showTopBar={showTopBar} />
-      <div className="text-contrast inset-shadow-glow mx-auto my-2 flex h-fit min-h-screen w-[95%] flex-col gap-4 rounded-xl px-4 pt-28 pb-4 backdrop-blur-xs">
+      {/* main content */}
+      <div className="text-contrast inset-shadow-glow dark:inset-shadow-glow/20 mx-auto my-2 flex h-fit min-h-screen w-[95%] flex-col gap-4 rounded-xl px-4 pt-28 pb-4 backdrop-blur-xs">
         <div className="after:inset-shadow-glow animate-in fade-in relative overflow-hidden rounded-2xl opacity-90 duration-400 after:absolute after:top-0 after:left-0 after:h-full after:w-full">
           <img
             src="/headshotcropped.png"
